Do not warn about ae on pipeline import
- Fix import issue
- Add tests in the existing test file only
