+++ tests/test_pipelines.py
@@ -1,7 +1,10 @@
-import pytest  # isort: skip
+"""
+Test pipeline component functionality. Also checks if warning is generated or not.
+For DropLowProbabilityItemPipeline, warning should not be generated.
+For AEPipeline, warning should be generated.
+"""
 
-scrapy = pytest.importorskip("scrapy")  # noqa
-
+import warnings
 from copy import deepcopy
 from unittest.mock import MagicMock, patch
 
@@ -16,6 +19,14 @@ from zyte_common_items import (
 )
 from zyte_common_items.base import ProbabilityMixin
 from zyte_common_items.pipelines import DropLowProbabilityItemPipeline
+
+import pytest  # isort: skip
+
+
+scrapy = pytest.importorskip("scrapy")  # noqa
+
+
+WARNING_MSG = r"The zyte_common_items.ae module is a temporary module"
 
 
 @pytest.mark.parametrize(
@@ -307,3 +318,37 @@ def test_get_item_name(item, expected_name):
     assert (
         DropLowProbabilityItemPipeline.get_item_name(MagicMock(), item) == expected_name
     )
+
+
+def test_warning():
+    """
+    Asserts if Deprication warning gets generated or not while
+    importing AEPipeline. It should generate a warning message.
+    """
+    warnings.filterwarnings("default")
+
+    with warnings.catch_warnings(record=True) as record:
+        from zyte_common_items.pipelines import AEPipeline
+
+        ae_pipeline = AEPipeline()
+    from zyte_common_items.ae import downgrade
+
+    warn_msg = str(record[0].message)
+    assert len(record) == 1
+    assert WARNING_MSG in warn_msg
+    assert ae_pipeline._downgrade == downgrade
+
+
+def test_no_warning():
+    """
+    Asserts if Deprication warning gets generated or not while
+    importing DropLowProbabilityItemPipeline. It should not generate a warning
+    message.
+    """
+    warnings.filterwarnings("default")
+    with warnings.catch_warnings(record=True) as record:
+        from zyte_common_items.pipelines import DropLowProbabilityItemPipeline
+
+        mock_crawler = MagicMock(spec=["spider", "stats"])
+        DropLowProbabilityItemPipeline(mock_crawler)
+    assert len(record) == 0
